Add authors to database (#21)
* Add models

* Add models to admin interface

* Setup memcached

* Setup initial new authors analysis

* Setup initial file hashing for deduplication

* Update new authors analysis

* Reverse getAuthorInfo

We no longer need getAuthorInfo. Furthermore, although we are no longe
using parseCSV, it is broken because getAuthorInfo is modified.

* Add auto analyzing of data if authenticated for authors

* Mark data to be removed after other files are stored to database

Data should not be sent in parsed form back and forth as it is more
efficient to handle it once server side. Furthermore, posting data
that is large can cause memory issues.

As an example: https://github.com/encode/django-rest-framework/issues/4760